Fix: rename wydr.js -> wdyr.js.
