+++ frontend/src/pages/_app.js
@@ -1,4 +1,4 @@
-import '../wydr'
+import '../wdyr'
 import React from 'react'
 
 import App from 'next/app'
